+++ airline-core/src/main/java/com/github/rvesse/airline/restrictions/factories/OptionRestrictionFactory.java
@@ -16,6 +16,7 @@
 package com.github.rvesse.airline.restrictions.factories;
 
 import java.lang.annotation.Annotation;
+import java.util.List;
 
 import com.github.rvesse.airline.restrictions.OptionRestriction;
 
@@ -34,4 +35,11 @@ public interface OptionRestrictionFactory {
      */
     public abstract OptionRestriction createOptionRestriction(Annotation annotation);
 
+    /**
+     * Gets a list of annotations that this factory can convert into option
+     * restrictions
+     * 
+     * @return List of supported annotations
+     */
+    public List<Class<? extends Annotation>> supportedOptionAnnotations();
 }