+++ airline-core/src/main/java/com/github/rvesse/airline/restrictions/factories/SimpleRestrictionsFactory.java
@@ -16,38 +16,44 @@
 package com.github.rvesse.airline.restrictions.factories;
 
 import java.lang.annotation.Annotation;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import com.github.rvesse.airline.annotations.restrictions.Port;
+import com.github.rvesse.airline.annotations.restrictions.Required;
+import com.github.rvesse.airline.annotations.restrictions.Unrestricted;
+import com.github.rvesse.airline.restrictions.AbstractCommonRestriction;
 import com.github.rvesse.airline.restrictions.ArgumentsRestriction;
+import com.github.rvesse.airline.restrictions.None;
 import com.github.rvesse.airline.restrictions.OptionRestriction;
-import com.github.rvesse.airline.restrictions.common.PortRestriction;
+import com.github.rvesse.airline.restrictions.common.IsRequiredRestriction;
 
-public class PortRestrictionFactory implements OptionRestrictionFactory, ArgumentsRestrictionFactory {
+public class SimpleRestrictionsFactory implements OptionRestrictionFactory, ArgumentsRestrictionFactory {
 
     @Override
     public ArgumentsRestriction createArgumentsRestriction(Annotation annotation) {
-        if (annotation instanceof Port) {
-            return createCommon((Port) annotation);
+        return createCommon(annotation);
+    }
+
+    private AbstractCommonRestriction createCommon(Annotation annotation) {
+        if (annotation instanceof Required) {
+            return new IsRequiredRestriction();
+        } else if (annotation instanceof Unrestricted) {
+            return new None();
         }
         return null;
     }
 
     @Override
     public OptionRestriction createOptionRestriction(Annotation annotation) {
-        if (annotation instanceof Port) {
-            return createCommon((Port) annotation);
-        }
-        return null;
+        return createCommon(annotation);
     }
 
-    protected final PortRestriction createCommon(Port annotation) {
-        return new PortRestriction(annotation.acceptablePorts());
-    }
-    
-    protected List<Class<? extends Annotation >> supportedAnnotations() {
-        return Collections.<Class<? extends Annotation>>singletonList(Port.class);
+    protected List<Class<? extends Annotation>> supportedAnnotations() {
+        List<Class<? extends Annotation>> supported = new ArrayList<>();
+        supported.add(Required.class);
+        supported.add(Unrestricted.class);
+        return supported;
     }
 
     @Override
@@ -59,4 +65,5 @@ public class PortRestrictionFactory implements OptionRestrictionFactory, Argumen
     public List<Class<? extends Annotation>> supportedOptionAnnotations() {
         return supportedAnnotations();
     }
+
 }